# chittyos-openai-gpt/openapi.yaml
openapi: 3.0.0
info:
  title: ChittyOS Infrastructure API for ChatGPT
  description: AI-driven legal tech infrastructure management with Cloudflare, blockchain, and compliance integrations
  version: 1.0.0
  contact:
    name: ChittyOS Support
    url: https://chitty.cc
    email: support@chitty.cc
servers:
  - url: https://api.chitty.cc/v1
    description: ChittyOS Production API
  - url: https://gateway.chitty.cc
    description: ChittyOS Gateway Worker
  - url: https://localhost:8000
    description: Local Development
paths:
...
  /legal/cases:
    post:
      operationId: createLegalCase
      summary: Create legal case
      description: Create a new legal case with full infrastructure provisioning
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - client_name
                - case_type
              properties:
                client_name:
                  type: string
                case_type:
                  type: string
                  enum: [litigation, contract, merger, compliance, intellectual_property]
                jurisdiction:
                  type: string
                priority:
                  type: string
                  enum: [low, normal, high, critical]
                auto_provision:
                  type: boolean
                  description: Automatically provision infrastructure
                  default: true
      responses:
        '200':
          description: Case created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CaseResponse'
...
components:
  schemas:
...
    CaseResponse:
      type: object
      properties:
        case_id:
          type: string
        chitty_id:
          type: string
        status:
          type: string
        infrastructure:
          type: object
          properties:
            worker:
              type: string
            database:
              type: string
            storage:
              type: string
...
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-ChittyOS-API-Key
    BearerAuth:
      type: http
      scheme: bearer
security:
  - ApiKeyAuth: []
  - BearerAuth: []